Ex2.14 Adding confirmation window

+++ part2/phonebook/src/components/PersonDisplay.jsx
@@ -1,7 +1,9 @@
 const PersonDisplay = ({person, deletePerson}) => {
     const deleteEntry = () =>{
-        console.log(`Delete '${person.name}'`)
-        deletePerson(person.id)
+        const confirmed  = window.confirm(`Delete '${person.name}'?`)
+        if (confirmed){
+            deletePerson(person.id)
+        }
     }
 
     return (
